refactor calendar grid to a shared component
